Refactoring folder structure pre-release
