+++ distiller/models/__init__.py
@@ -18,8 +18,12 @@
 
 import torch
 import torchvision.models as torch_models
+import os
+import sys
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 import models.cifar10 as cifar10_models
 import models.imagenet as imagenet_extra_models
+sys.path.pop(0)
 
 import logging
 msglogger = logging.getLogger()
