+++ tests/common.py
@@ -21,7 +21,7 @@ module_path = os.path.abspath(os.path.join('..'))
 if module_path not in sys.path:
     sys.path.append(module_path)
 import distiller
-from models import create_model
+from distiller.models import create_model
 
 
 def setup_test(arch, dataset, parallel):
