+++ tests/test_basic.py
@@ -22,7 +22,7 @@ module_path = os.path.abspath(os.path.join('..'))
 if module_path not in sys.path:
     sys.path.append(module_path)
 import distiller
-import models
+import distiller.models
 
 
 def test_sparsity():
