+++ tests/test_pruning.py
@@ -27,8 +27,8 @@ except ImportError:
     import distiller
 import common
 import pytest
-from models import create_model
-from apputils import save_checkpoint, load_checkpoint
+from distiller.models import create_model
+from distiller.apputils import save_checkpoint, load_checkpoint
 
 # Logging configuration
 logging.basicConfig(level=logging.INFO)
